Adding support for Rails 6 and updating test dependencies
The lastest supported and tested version for attributes_sanitizer was Rails 5. In this commit I'm adding support to Rails 6 excluding 7+ as it is about to be released.

I've also taken the opportunity to improve test dependencies.
Rails 5.2 relies on bootsnap to speed up it's execution time.

And I've  added .ruby-version file to .gitignore.